refactor: doc index structure
