+++ docarray/doc_index/__init__.py
@@ -1,0 +1,1 @@
+from docarray.doc_index.backends.hnswlib import HnswDocumentIndex
