+++ tests/integrations/doc_index/doc_index/hnswlib/test_find.py
@@ -3,7 +3,7 @@ import pytest
 from pydantic import Field
 
 from docarray import BaseDocument
-from docarray.doc_index.backends.hnswlib_doc_index import HnswDocumentIndex
+from docarray.doc_index import HnswDocumentIndex
 from docarray.typing import NdArray
 
 pytestmark = [pytest.mark.slow, pytest.mark.doc_index]
